release: fix reboot check
The reboot check never finished, because `progress.finished` was still
waiting for a parent task to finish that would only succeed *after* the
call towards `wait_for_vm_reboot` had returned.
In general, that function only needs to be interested in the finishing
of its *own* task, not in the overall progress of all – potentially
unrelated – tasks.